Fix ID generation
As files are stored as their handlers in the file system and
Windows' case insensitive it's simpler to make ID generation
work in a case insensitive way than changing how the server
stores uploaded files in the file system.

BREAKING CHANGE if the server was used in a way that at least
36 IDs were generated. If less IDs were generated it's not
a breaking change because the Base32 alphabet used is compatible
with Base64 for the first generated IDs.

Also now the counter used for the default ID generation is stored
in the server state avoiding bugs when restarting the server.